requirements.txt
django==4.2.24
django-split-settings==1.3.2
flake8==7.3.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
pytest==8.4.2